python-scraper/requirements.txt
fastapi
uvicorn
playwright
google-genai
pydantic
